Fix handling of trailing commas in associative arrays
A trailing comma results in an ArrayItem with nil Key and Val. The nil Key was
triggering false positives in mixedArrayKeys checks of associative arrays
because it assumed there was a value with an implicit key. This now checks that
Val is non-nil before treating it as having an implicit key.