Fix handling of variadic parameters and arguments
A variadic parameter no longer increases the minimum argument count for a
function, since it is valid not to pass any arguments for the variadic
parameter, in which case it is initialized to an empty array.

Passing an array with a "..." prefix as the last argument in a function call
now assumes the array has sufficient values to match the minimum argument count
for the called function.