Do not treat variable assignments in argument lists as unused variables
If a function is called as "foo($min = 1, $max = 10)", then $min and $max
should not be added to the list of initially unused variables. This is a
convention used to clarify the meaning of arguments when it is not otherwise
clear from the context, such as functions that take multiple boolean arguments.