Treat aliased variables as non-local
When a variable reference is assigned to another variable, it is difficult to
track if a write to one is unused because it can be read by another, and
references can be used to escape the local context. In order to avoid false
positives for warnings about unused variables, it's better to treat the left
and right hand variables as non-local.